Update
TCPFighter MMO
 - 로직 업데이트 방식 변경 (프레임 방식 -> 델타 타임 방식)

+++ TCPFighter MMO/MMO_FightServer/MMO_FightServer/Object.h
@@ -32,8 +32,8 @@ struct CHARACTER
 	unsigned long action;
 	unsigned char direction;
 	unsigned char hp;
-	unsigned short x;
-	unsigned short y;
+	float x;
+	float y;
 	SECTOR curSector;
 	SECTOR oldSector;
 };
